pyproject.toml
[tool.poetry]
name = "winston"
version = "0.1.0"
description = "Winston is the silence monitor for University Radio Nottingham."
authors = ["Joe Banks <joe.banks@urn1350.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
python-dotenv = "^0.21.0"
loguru = "^0.6.0"
deepmerge = "^1.1.0"
httpx = "^0.23.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
